src/backend/requirements.txt
flask
python-dotenv
requests
flask-cors
pycristoforo
numpy
Shapely